move fault_injection_test.go to toplevel trafficmanagement package to avoid slows down the test
