+++ tests/integration/istioio/trafficmanagement/scripts/faultinjection/virtual_service_abort_creation.txt
