+++ tests/integration/istioio/trafficmanagement/scripts/faultinjection/virtual_service_delay_creation.txt
